Replace XMLHttpRequest with Fetch API
This PR replaces the synchronous XMLHttpRequest calls with the more modern and asynchronous Fetch API.
Where possible, requests are started in parallel. The execution only waits for the execution once we need the results to complete.
This PR is only a suggestion to modernize the code a little and provide some minor improvements in responsiveness.