Add frontend for migration to React + Typescript (#140)
* Initial draft for react build

This commit adds the general structure to use React with Vite for the escape room frontend.
It adds a build script and CI checks for GitHub PRs.
This PR does NOT provide a rewritten frontend in React and does not replace the current frontend in the escape room (to keep it working until the migration is done)

* Update kind version to fix bug

kind load docker-image was broken (see [here](https://github.com/kubernetes-sigs/kind/issues/3853)).

* Fix rewrite-target rule for frontend